Changed long to int32_t throughout
git-svn-id: svn://orca.physics.unc.edu/Orca64@11937 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/App Delegate/ORAppDelegate.h
@@ -36,7 +36,7 @@
 	NSString* ethernetHardwareAddress;
 	BOOL configLoadedOK;
 	NSOperationQueue* queue;
-	unsigned long heartbeatCount;
+	uint32_t heartbeatCount;
 	BOOL delayTermination;
 }
 
@@ -101,9 +101,9 @@ extern NSString* OROrcaFinalQuitNotice;
 
 @interface ORHeartBeatOp : NSOperation
 {
-	unsigned long heartbeatCount;
+	uint32_t heartbeatCount;
 }
-- (id) init:(unsigned long)aCount;
+- (id) init:(uint32_t)aCount;
 - (void) main;
 @end
