+++ Source/Experiments/Ncd/CableCheckTask/NcdCableCheckTask.h
@@ -58,15 +58,15 @@
     BOOL passed;
     ORShaperModel* currentShaper;
     int currentShaperChannel;
-    unsigned long startingShaperCounts;
+    uint32_t startingShaperCounts;
     
     NcdMuxBoxModel* currentMux;
     int currentMuxChannel;
-    unsigned long startingMuxCounts;
+    uint32_t startingMuxCounts;
 
     NSArray* currentScopes;
     int currentScopeChannel;
-    unsigned long startingScopeCounts;
+    uint32_t startingScopeCounts;
 
     unsigned short oldMuxThreshold[13];
     unsigned short muxThreshold;
